refactor: utils renaming and relocation
